chore: split common SDL `plat_*` functions to separate module
